🧪 Add comprehensive mock tests for setup.py
